# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SiapPA API",
        "description": "API documentation for SiapPA application",
        "contact": {
            "email": "support@siappa.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/auth/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "description": "Returns a list of all users",
                "operationId": "getAllUsers",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "description": "Creates a new user and returns an authentication token",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nama pengguna",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Email pengguna",
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "description": "Kata sandi pengguna",
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "description": "Authenticates a user and returns an authentication token",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email pengguna",
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "description": "Kata sandi pengguna",
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "description": "Logs out the authenticated user and revokes the token",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Pesan sukses",
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get authenticated user profile",
                "description": "Returns the profile of the authenticated user",
                "operationId": "getUserProfile",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Status keberhasilan",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "message": {
                                            "description": "Pesan sukses",
                                            "type": "string",
                                            "example": "User profile retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update authenticated user profile",
                "description": "Updates the profile of the authenticated user",
                "operationId": "updateProfile",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nama pengguna",
                                        "type": "string",
                                        "example": "John Doe Updated",
                                        "nullable": true
                                    },
                                    "foto_profil": {
                                        "description": "Path atau URL foto profil",
                                        "type": "string",
                                        "example": "path/to/newfoto.jpg",
                                        "nullable": true
                                    },
                                    "password": {
                                        "description": "Kata sandi baru",
                                        "type": "string",
                                        "example": "newpassword123",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Status keberhasilan",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Pesan sukses",
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "description": "ID unik pengguna",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nama pengguna",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Email pengguna",
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "foto_profile": {
                        "description": "Path atau URL foto profil",
                        "type": "string",
                        "example": "path/to/foto.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Waktu pembuatan",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-21T20:40:00Z"
                    },
                    "updated_at": {
                        "description": "Waktu pembaruan",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-21T20:40:00Z"
                    }
                },
                "type": "object"
            },
            "AuthResponse": {
                "properties": {
                    "token": {
                        "description": "Token autentikasi",
                        "type": "string",
                        "example": "your-auth-token"
                    },
                    "token_type": {
                        "description": "Jenis token",
                        "type": "string",
                        "example": "Bearer"
                    },
                    "user_data": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "nullable": true,
                        "description": "Data pengguna"
                    },
                    "success": {
                        "description": "Status keberhasilan",
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    },
                    "message": {
                        "description": "Pesan sukses",
                        "type": "string",
                        "example": "Login successful",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "description": "Pesan error",
                        "type": "string",
                        "example": "Login failed, please check your credentials."
                    },
                    "success": {
                        "description": "Status keberhasilan",
                        "type": "boolean",
                        "example": false,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "error": {
                        "description": "Kesalahan validasi",
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}